Fixing up unit tests
